requirements.txt
transformers
pillow
requests
tqdm
gitpython
huggingface_hub
bitsandbytes
